backend/requirements.txt
Flask==2.3.2
easyocr==1.6.2
torch==2.1.0
torchvision==0.16.0
Pillow==10.2.0
Werkzeug==2.3.6
